+++ EF4.java
@@ -3,267 +3,83 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package ef3;
 
-import java.util.Scanner;
+import javax.swing.JFrame;
+import javax.swing.JButton;
+import java.awt.event.ActionListener;
+import java.awt.event.ActionEvent;
+import javax.swing.JOptionPane;
+import javax.swing.JTextField;
+import javax.swing.JLabel;
 
 /**
  *
  * @author LuisMiguel
  */
-public class EF4 {
-    
-    Scanner scanner;
-        int op;
+public class EF4 extends JFrame implements ActionListener {
 
-        public EF4() {
-            scanner = new Scanner(System.in);
-            try{
-                menu();
-            }catch(Exception e){
-                System.out.println("Ingresa valores validos porfavor");
-                new EF4();
+    /**
+     * @param args the command line arguments
+     */
+    JFrame ef4;
+    JButton boton;
+    JTextField text;
+    JLabel etiqueta;
+
+    public EF4() {
+
+        ef4 = new JFrame("EF4");
+        ef4.setSize(640, 260);
+        ef4.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        ef4.setVisible(true);
+        ef4.setResizable(false);
+        ef4.setLocationRelativeTo(null);
+        ef4.setLayout(null);
+
+        boton = new JButton("Sumar split");
+        boton.setBounds(270, 150, 100, 30);
+        ef4.add(boton);
+        boton.addActionListener(this);
+
+        text = new JTextField();
+        text.setBounds(95, 60, 450, 30);
+        ef4.add(text);
+        
+        etiqueta = new JLabel();
+        etiqueta.setText("Ingrese numeros separados por comas para sumatoria");
+        etiqueta.setBounds(60, 100, 500, 30);
+        ef4.add(etiqueta);
+    }
+
+    @Override
+    public void actionPerformed(ActionEvent e) {
+        if (e.getSource() == boton) {
+            etiqueta.setText(String.valueOf(sumar()));
+        }
+    }
+
+    public int sumar() {
+        int i;
+        int sumatoria = 0;
+        try {
+            String[] texto = text.getText().split(",");
+            int[] numeros = new int[texto.length];
+
+            for (i = 0; i < texto.length; i++) {
+                numeros[i] = Integer.parseInt(texto[i]);
+            }
+            for (i = 0; i < numeros.length; i++) {
+                sumatoria = sumatoria + numeros[i];
             }
             
+        } catch (Exception e) {
+            JOptionPane.showMessageDialog(text, e);
         }
-
-        private void menu() {
-
-            String mensaje;
-        while (op != 3) {
-            System.out.println("CRIPTOGRAFIA \n"
-                    + "1.) Cifrar mensaje. \n"
-                    + "2.) Descifrar mensaje. \n"
-                    + "3.) Salir");
-            op = scanner.nextInt();
-            scanner.nextLine();
-
-            switch (op) {
-                case 1:
-                    System.out.println("Ingrese el mensaje a cifrar");
-                    mensaje = scanner.nextLine();
-                    System.out.println(cifrar(mensaje));
-                    break;
-                case 2:
-                    System.out.println("Ingrese el mensaje a descifrar");
-                    mensaje = scanner.nextLine();
-                    System.out.println(descifrar(mensaje));
-
-                    break;
-                case 3:
-                    System.out.println("Adios!");
-                    break;
-                default:
-                    System.out.println("Debes ingresar un valor valido");
-                    break;
-            }
-        }
-
-    }
-        
-        private int letrastonumeros(char entrada) {
-
-            //nos devuelve el numero en el alfabeto de la letra, para espacio en blanco es 0
-        int salida = 0;
-        switch (entrada) {
-            case 'a':
-                salida = 1;
-                break;
-            case 'b':
-                salida = 2;
-                break;
-            case 'c':
-                salida = 3;
-                break;
-            case 'd':
-                salida = 4;
-                break;
-            case 'e':
-                salida = 5;
-                break;
-            case 'f':
-                salida = 6;
-                break;
-            case 'g':
-                salida = 7;
-                break;
-            case 'h':
-                salida = 8;
-                break;
-            case 'i':
-                salida = 9;
-                break;
-            case 'j':
-                salida = 10;
-                break;
-            case 'k':
-                salida = 11;
-                break;
-            case 'l':
-                salida = 12;
-                break;
-            case 'm':
-                salida = 13;
-                break;
-            case 'n':
-                salida = 14;
-                break;
-            case 'o':
-                salida = 15;
-                break;
-            case 'p':
-                salida = 16;
-                break;
-            case 'q':
-                salida = 17;
-                break;
-            case 'r':
-                salida = 18;
-                break;
-            case 's':
-                salida = 19;
-                break;
-            case 't':
-                salida = 20;
-                break;
-            case 'u':
-                salida = 21;
-                break;
-            case 'v':
-                salida = 22;
-                break;
-            case 'w':
-                salida = 23;
-                break;
-            case 'x':
-                salida = 24;
-                break;
-            case 'y':
-                salida = 25;
-                break;
-            case 'z':
-                salida = 26;
-                break;
-            case ' ': //para espacios en blanco
-                salida = 0;
-                break;
-
-        }
-        return salida;
-    }
-        private char numerostoletras(int entrada){
-            char salida = ' ';
-            switch (entrada) {
-                case 1:
-                    salida = 'a';
-                    break;
-                case 2:
-                    salida = 'b';
-                    break;
-                case 3:
-                    salida = 'c';
-                    break;
-                case 4:
-                    salida = 'd';
-                    break;
-                case 5:
-                    salida = 'e';
-                    break;
-                case 6:
-                    salida = 'f';
-                    break;
-                case 7:
-                    salida = 'g';
-                    break;
-                case 8:
-                    salida = 'h';
-                    break;
-                case 9:
-                    salida = 'i';
-                    break;
-                case 10:
-                    salida = 'j';
-                    break;
-                case 11:
-                    salida = 'k';
-                    break;
-                case 12:
-                    salida = 'l';
-                    break;
-                case 13:
-                    salida = 'm';
-                    break;
-                case 14:
-                    salida = 'n';
-                    break;
-                case 15:
-                    salida = 'o';
-                    break;
-                case 16:
-                    salida = 'p';
-                    break;
-                case 17:
-                    salida = 'q';
-                    break;
-                case 18:
-                    salida = 'r';
-                    break;
-                case 19:
-                    salida = 's';
-                    break;
-                case 20:
-                    salida = 't';
-                    break;
-                case 21:
-                    salida = 'u';
-                    break;
-                case 22:
-                    salida = 'v';
-                    break;
-                case 23:
-                    salida = 'w';
-                    break;
-                case 24:
-                    salida = 'x';
-                    break;
-                case 25:
-                    salida = 'y';
-                    break;
-                case 26:
-                    salida = 'z';
-                    break;
-                case 0: //para espacios en blanco
-                    salida = ' ';
-                    break;
-            }
-            return salida;
-        }
-    
-    public String cifrar(String mensaje) {
-
-        int i;
-        char[] aux = mensaje.toCharArray();
-
-        for (i = 0; i < aux.length; i++) {
-            //(5X + 11 ) mod 26 + 1
-            aux[i] = numerostoletras((5 * letrastonumeros(aux[i]) + 11) % 26 + 1);
-        }
-        mensaje = String.valueOf(aux);
-
-        return mensaje;
+        return sumatoria;
     }
 
-    public String descifrar(String mensaje) {
-        int i;
-        char[] aux = mensaje.toCharArray();
-
-        for (i = 0; i < aux.length; i++) {
-            //(ai * (C - 11)) mod 26
-            aux[i] = numerostoletras((21) * (letrastonumeros(aux[i]) - 11) % 26);
-        }
-        mensaje = String.valueOf(aux);
-
-        return mensaje;
+    public static void main(String[] args) {
+        new EF4();
     }
+
 }
-
